backend/requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Network and HTTP
requests==2.31.0

# System monitoring
psutil==5.9.6

# File processing
PyPDF2==3.0.1
Pillow==10.1.0
python-magic==0.4.27
PyMuPDF==1.23.8

# OCR dependencies
paddlepaddle==2.5.2
paddleocr==2.7.0.3
opencv-python==4.8.1.78
numpy==1.24.3
beautifulsoup4==4.12.2

# Character encoding
chardet==5.2.0

# Data validation and serialization
pydantic==2.5.0
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.92.1

# Development
python-dotenv==1.0.0